# applications/visualizer/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "C. Elegans Visualizer",
    "version": "1.0.0",
    "description": ""
  },
  "paths": {
    "/api/datasets": {
      "get": {
        "operationId": "get_datasets",
        "summary": "Get Datasets",
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ids"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Dataset"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            }
          }
        },
        "description": "Returns all datasets or a filtered list based on provided IDs",
        "tags": [
          "datasets"
        ]
      }
    },
...
    "/api/datasets/{dataset}": {
      "get": {
        "operationId": "get_dataset",
        "summary": "Get Dataset",
        "parameters": [
          {
            "in": "path",
            "name": "dataset",
            "schema": {
              "title": "Dataset",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dataset"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "description": "Returns a specific dataset",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/datasets/{dataset}/neurons": {
      "get": {
        "operationId": "get_dataset_neurons",
        "summary": "Get Dataset Neurons",
        "parameters": [
          {
            "in": "path",
            "name": "dataset",
            "schema": {
              "title": "Dataset",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Neuron"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "description": "Returns all the neurons of a dedicated dataset",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/cells/search": {
      "get": {
        "operationId": "search_cells",
        "summary": "Search Cells",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "dataset_ids",
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dataset Ids"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Neuron"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": [
          "neurons"
        ]
      }
    },
    "/api/cells": {
      "get": {
        "operationId": "get_all_cells",
        "summary": "Get All Cells",
        "parameters": [
          {
            "in": "query",
            "name": "dataset_ids",
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dataset Ids"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedNeuron"
                }
              }
            }
          }
        },
        "description": "Returns all the cells (neurons) from the DB",
        "tags": [
          "neurons"
        ]
      }
    },
...
    "/api/connections": {
      "get": {
        "operationId": "get_connections",
        "summary": "Get Connections",
        "parameters": [
          {
            "in": "query",
            "name": "cells",
            "schema": {
              "title": "Cells",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "dataset_ids",
            "schema": {
              "title": "Dataset Ids",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "dataset_type",
            "schema": {
              "title": "Dataset Type",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "threshold_chemical",
            "schema": {
              "default": 3,
              "title": "Threshold Chemical",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "threshold_electrical",
            "schema": {
              "default": 3,
              "title": "Threshold Electrical",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "include_neighboring_cells",
            "schema": {
              "default": false,
              "title": "Include Neighboring Cells",
              "type": "boolean"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "include_annotations",
            "schema": {
              "default": false,
              "title": "Include Annotations",
              "type": "boolean"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Connection"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            }
          }
        },
        "description": "Gets the connections of a dedicated Dataset",
        "tags": [
          "connectivity"
        ]
      }
    },
    "/api/connections/{datasetId}/download": {
      "get": {
        "operationId": "get_dataset_connectivity",
        "summary": "Get Dataset Connectivity",
        "parameters": [
          {
            "in": "path",
            "name": "datasetId",
            "schema": {
              "title": "Datasetid",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "default": "csv",
              "enum": [
                "csv",
                "json"
              ],
              "title": "Format",
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Download the connections of a dedicated Dataset in either CSV or JSON format (default CSV).",
        "tags": [
          "connectivity"
        ]
      }
    },
    "/api/connections/{datasetId}": {
      "get": {
        "operationId": "get_dataset_connections",
        "summary": "Get Dataset Connections",
        "parameters": [
          {
            "in": "path",
            "name": "datasetId",
            "schema": {
              "title": "Datasetid",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "exclude_class",
            "schema": {
              "default": false,
              "title": "Exclude Class",
              "type": "boolean"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "neurons",
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Neurons"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "items": {
                        "$ref": "#/components/schemas/RawConnection"
                      },
                      "type": "array"
                    },
                    {
                      "items": {
                        "$ref": "#/components/schemas/GroupedConnection"
                      },
                      "type": "array"
                    }
                  ],
                  "title": "Response"
                }
              }
            }
          }
        },
        "description": "Gets the connections of a dedicated Dataset\nConnections includes connection towards the neurons and their classes by default.\nif exclude_class is set to true: the neuron classes (higher level neuron) is not included.",
        "tags": [
          "connectivity"
        ]
      }
    },
    "/api/synapses": {
      "get": {
        "operationId": "get_dataset_synapses",
        "summary": "Get Dataset Synapses",
        "parameters": [
          {
            "in": "query",
            "name": "datasetIds",
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Datasetids",
              "type": "array"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "neurons",
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Neurons",
              "type": "array"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupedSynapse"
                }
              }
            }
          }
        },
        "tags": [
          "synapses"
        ]
      }
    },
    "/api/populate_db": {
      "get": {
        "operationId": "populate_db",
        "summary": "Populate Db",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "ingestion"
        ],
        "security": [
          {
            "BasicAuthSuperUser": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Dataset": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "neuron3DUrl": {
            "title": "Neuron3Durl",
            "type": "string"
          },
          "emData": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EMData"
              },
              {
                "type": "null"
              }
            ]
          },
          "collection": {
            "maxLength": 20,
            "title": "Collection",
            "type": "string"
          },
          "name": {
            "maxLength": 50,
            "title": "Name",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "time": {
            "title": "Time",
            "type": "number"
          },
          "visualTime": {
            "title": "Visual Time",
            "type": "number"
          },
          "type": {
            "maxLength": 50,
            "title": "Type",
            "type": "string"
          },
          "axes": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Axes"
          }
        },
        "required": [
          "id",
          "neuron3DUrl",
          "emData",
          "collection",
          "name",
          "description",
          "time",
          "visualTime",
          "type"
        ],
        "title": "Dataset",
        "type": "object"
      },
...
      "ErrorMessage": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorMessage",
        "type": "object"
      },
      "Neuron": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "datasetIds": {
            "items": {
              "type": "string"
            },
            "title": "Datasetids",
            "type": "array"
          },
          "model3DUrls": {
            "items": {
              "type": "string"
            },
            "title": "Model3Durls",
            "type": "array"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "nclass": {
            "maxLength": 30,
            "title": "Nclass",
            "type": "string"
          },
          "neurotransmitter": {
            "maxLength": 10,
            "title": "Neurotransmitter",
            "type": "string"
          },
          "type": {
            "maxLength": 10,
            "title": "Type",
            "type": "string"
          },
          "embryonic": {
            "default": false,
            "title": "Embryonic",
            "type": "boolean"
          },
          "inhead": {
            "default": false,
            "title": "Inhead",
            "type": "boolean"
          },
          "intail": {
            "default": false,
            "title": "Intail",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "datasetIds",
          "model3DUrls",
          "reference",
          "nclass",
          "neurotransmitter",
          "type"
        ],
        "title": "Neuron",
        "type": "object"
      },
...
      "PagedNeuron": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Neuron"
            },
            "title": "Items",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "PagedNeuron",
        "type": "object"
      },
      "Connection": {
        "properties": {
          "annotations": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Annotations",
            "type": "array"
          },
          "synapses": {
            "additionalProperties": {
              "type": "integer"
            },
            "default": {},
            "title": "Synapses",
            "type": "object"
          },
          "pre": {
            "maxLength": 30,
            "title": "Pre",
            "type": "string"
          },
          "post": {
            "maxLength": 30,
            "title": "Post",
            "type": "string"
          },
          "type": {
            "maxLength": 20,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "pre",
          "post",
          "type"
        ],
        "title": "Connection",
        "type": "object"
      },
...
      "RawConnection": {
        "properties": {
          "pre": {
            "maxLength": 30,
            "title": "Pre",
            "type": "string"
          },
          "post": {
            "maxLength": 30,
            "title": "Post",
            "type": "string"
          },
          "type": {
            "maxLength": 20,
            "title": "Type",
            "type": "string"
          },
          "synapses": {
            "title": "Synapses",
            "type": "integer"
          }
        },
        "required": [
          "pre",
          "post",
          "type",
          "synapses"
        ],
        "title": "RawConnection",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BasicAuthSuperUser": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "servers": []
}